refactor + add serial fb algorithms
